requirements.txt
gradio
numpy
opencv_python
opencv_python_headless
Pillow
PyYAML
torch
torchvision
transformers
